--- a/src/arm64/simd.h
+++ b/src/arm64/simd.h
@@ -3,6 +3,7 @@
 
 #include "simdjson.h"
 #include "simdprune_tables.h"
+#include "arm64/bitmanipulation.h"
 #include "arm64/intrinsics.h"
 
 namespace simdjson::arm64::simd {