… benchmark/minifiercompetition.cpp
… benchmark/parse.cpp
… include/CMakeLists.txt
… include/simdjson/jsonminifier.h
… src/CMakeLists.txt
… src/generic/json_structural_indexer.h
… src/generic/stage1_find_marks.h
… src/haswell/bitmanipulation.h
… src/haswell/implementation.h
… src/haswell/simd.h
… src/haswell/stage1_find_marks.h
… src/implementation.cpp
… src/jsonminifier.cpp
… src/simdjson.cpp
… src/westmere/bitmanipulation.h
… src/westmere/implementation.h
… src/westmere/simd.h
… src/westmere/stage1_find_marks.h
… tests/basictests.cpp
… tools/minify.cpp