On @jkeiser's new branch (jkeiser/minify) we use 2.15 cycles per input byte to minify...

```
$ git checkout jkeiser/minify
$ make minifiercompetition && ./minifiercompetition jsonexamples/twitter.json 
json_minify                                     :    4.189 cycles per input byte (best)    4.196 cycles per input byte (avg)    0.813 GB/s (error margin: 0.001 GB/s)
simdjson->minify                                :    2.152 cycles per input byte (best)    2.171 cycles per input byte (avg)    1.582 GB/s (error margin: 0.014 GB/s)
minisize = 466906, original size = 631515  (minified down to 73.93 percent of original) 
```

On this new PR (which is on top of jkeiser/minify), we get 0.6 cycles per input byte to minify...

```
$ git checkout dlemire/fastminifier
$ make minifiercompetition && ./minifiercompetition jsonexamples/twitter.json 
json_minify                                     :    4.191 cycles per input byte (best)    4.203 cycles per input byte (avg)    0.813 GB/s (error margin: 0.002 GB/s)
simdjson->minify                                :    0.609 cycles per input byte (best)    0.612 cycles per input byte (avg)    5.576 GB/s (error margin: 0.026 GB/s)
minisize = 466906, original size = 631515  (minified down to 73.93 percent of original) 
```

Of course, results will depend on the dataset and the processor.

This is not yet well tested, it needs to reviewed, but it should be essentially correct.

This would fix https://github.com/lemire/simdjson/issues/384