@lemire 2020-03-20
Credit-wise, this PR is *on top* of jkeiser/minify.