@lemire 2020-03-20
(I don't think speeding up the minifier was critical work. The real motivation here is to test out the new design that @jkeiser pushed.)