@lemire 2020-03-20
We can now see the stats with all possible implementations...

```
$ ./minifiercompetition jsonexamples/twitter.json 
json_minify                                     :    4.188 cycles per input byte (best)    4.202 cycles per input byte (avg)    0.813 GB/s (error margin: 0.003 GB/s)
simdjson->minify+haswell                        :    0.600 cycles per input byte (best)    0.602 cycles per input byte (avg)    5.662 GB/s (error margin: 0.019 GB/s)
simdjson->minify+westmere                       :    0.787 cycles per input byte (best)    0.788 cycles per input byte (avg)    4.319 GB/s (error margin: 0.006 GB/s)
simdjson->minify+fallback                       :    4.182 cycles per input byte (best)    4.197 cycles per input byte (avg)    0.814 GB/s (error margin: 0.003 GB/s)
```

This is pretty much the kind of result you want.

I also made it so that the minifier tool can take a flag so that you can specify the architecture you want. This might allow someone to sidestep AVX if they so desire.