@lemire 2020-03-20
@jkeiser wrote "I bet we could pull 5GB/s". Indeed we can. In fact, we achieve "5.662 GB/s" on the twitter dataset on my test above.