@jkeiser 2020-03-20
> Note that this problem becomes trivial (and likely very fast) if you assume AVX-512.

Big time. Taking this algorithm and turning it into something we can reuse will also let us vectorize other things as well, such as copying string bytes.

> I also made it so that the minifier tool can take a flag so that you can specify the architecture you want. This might allow someone to sidestep AVX if they so desire.

I think we should add an environment variable, SIMDJSON_IMPLEMENTATION=<name>, and use that if it's set. That way we don't have to make arguments for all our programs.