@jkeiser 2020-03-20
> (I don't think speeding up the minifier was critical work. The real motivation here is to test out the new design that @jkeiser pushed.)

Also agreed. This is sort of a toy problem, and we could likely make it faster too--the current implementation spends some extra instructions finding operators ([]{},:), which is unnecessary for minification. But I don't think we care.