@jkeiser 2020-03-20
Oh, thank you for graciously not commenting on my rename of "hamming." I don't know who he is and it's always confused me :)