@lemire 2020-03-20
https://en.wikipedia.org/wiki/Hamming_weight