@lemire 2020-03-20
> the current implementation spends some extra instructions finding operators ([]{},:), which is unnecessary for minification. But I don't think we care.

I don't care per se, but are we sure that it does so?