@lemire 2020-03-20
This PR:
```
simdjson->minify+haswell                	:    0.474 cycles per input byte (best)    0.483 cycles per input byte (avg)    6.510 GB/s (error margin: 0.116 GB/s)
```

Version of this PR where I destroy the classification code:

```
$ git diff
diff --git a/src/haswell/stage1_find_marks.h b/src/haswell/stage1_find_marks.h
index 6850bb3..696af0c 100644
--- a/src/haswell/stage1_find_marks.h
+++ b/src/haswell/stage1_find_marks.h
@@ -34,10 +34,11 @@ really_inline json_character_block json_character_block::classify(const simd::si
     return _in == simd8<uint8_t>(_mm256_shuffle_epi8(whitespace_table, _in));
   }).to_bitmask();

-  uint64_t op = in.map([&](simd8<uint8_t> _in) {
+  uint64_t op = 0;
+// in.map([&](simd8<uint8_t> _in) {
     // | 32 handles the fact that { } and [ ] are exactly 32 bytes apart
-    return (_in | 32) == simd8<uint8_t>(_mm256_shuffle_epi8(op_table, _in-','));
-  }).to_bitmask();
+  //  return (_in | 32) == simd8<uint8_t>(_mm256_shuffle_epi8(op_table, _in-','));
+ // }).to_bitmask();
   return { whitespace, op };
```

I get...

```
simdjson->minify+haswell                	:    0.475 cycles per input byte (best)    0.582 cycles per input byte (avg)    6.510 GB/s (error margin: 1.210 GB/s)
```


@jkeiser What makes you think that we pay a price of the identification of the structural character? My guess is that with all the inlining, the compiler figures out that it does not need to do this work and does away with it.

Do you have evidence to the contrary?