@lemire 2020-03-20
@jkeiser Honestly, I am a bit confused, because I thought that the selling point of your new approach was that I could basically work on top of it and only use what I need without paying for the rest (either in code complexity or performance). That's what I think it does... if you think differently, I'd like to know!