@jkeiser 2020-03-20
> > the current implementation spends some extra instructions finding operators ([]{},:), which is unnecessary for minification. But I don't think we care.
> 
> I don't care per se, but are we sure that it does so?

On ARM, definitely: we recognize 64 bytes of operators and whitespace in one inseparable mask reduction. Instead, we could find 128 bytes of whitespace just as efficiently. Again, I super don't care and don't think we should fix it: it only applies to minification.

On Intel, I am hoping the optimizer notices it's not used and removes it. (I suspect it does.)