@jkeiser 2020-03-20
> @jkeiser Honestly, I am a bit confused, because I thought that the selling point of your new approach was that I could basically work on top of it and only use what I need without paying for the rest (either in code complexity or performance). That's what I think it does... if you think differently, I'd like to know!

Yep. I believe it does, particularly by relying on the optimizer. I really, really don't mean to imply that we need a lower level. Everything *except* minification needs every single bit of code here (and then some!).

I think of it as a token scanner: it will tell you where tokens are and tell you a lot about what characters are inside each one. The biggest point of the approach is that it lets you mix and match how you *use* the tokens. Whether you validate strings, UTF-8, or structure. Even whether/how you actually iterate the tokens, is left to the caller, which is the most important bit for performance: it doesn't attempt to shift cardinality from bytes to tokens. That's where you start getting branchy.