@lemire 2020-03-20
> On ARM, definitely: we recognize 32 bytes of operators and whitespace in one inseparable mask reduction. Instead, we could find 64 bytes of whitespace just as efficiently. Again, I super don't care and don't think we should fix it: it only applies to minification.

I added a note to this effect in the code. I did not remember this difference between x64 and NEON. Still. The performance is good on ARM. We are still minifying at the speed of stage 1. It is hard to beat by a lot.

Because I left a note, if someone cares, there is the documentation and the opportunity to optimize it further. I'll open an issue.