@lemire 2020-03-20
Merging.